[bcs] split bcs SDK into multiple files
